+++ lib/core-common/src/index.ts
@@ -11,6 +11,7 @@ export * from './utils/load-custom-presets';
 export * from './utils/load-main-config';
 export * from './utils/get-storybook-configuration';
 export * from './utils/get-storybook-info';
+export * from './utils/get-storybook-refs';
 export * from './utils/load-manager-or-addons-file';
 export * from './utils/load-preview-or-config-file';
 export * from './utils/log-config';