+++ lib/core-common/src/presets.ts
@@ -1,4 +1,4 @@
-import dedent from 'ts-dedent';
+import { dedent } from 'ts-dedent';
 import { logger } from '@storybook/node-logger';
 import {
   CLIOptions,
@@ -10,6 +10,7 @@ import {
 } from './types';
 import { loadCustomPresets } from './utils/load-custom-presets';
 import { safeResolve, safeResolveFrom } from './utils/safeResolve';
+import { interopRequireDefault } from './utils/interpret-require';
 
 const isObject = (val: unknown): val is Record<string, any> =>
   val != null && typeof val === 'object' && Array.isArray(val) === false;
@@ -153,17 +154,7 @@ const map =
     return undefined;
   };
 
-function interopRequireDefault(filePath: string) {
-  // eslint-disable-next-line global-require,import/no-dynamic-require
-  const result = require(filePath);
-
-  const isES6DefaultExported =
-    typeof result === 'object' && result !== null && typeof result.default !== 'undefined';
-
-  return isES6DefaultExported ? result.default : result;
-}
-
-function getContent(input: any) {
+async function getContent(input: any) {
   if (input.type === 'virtual') {
     const { type, name, ...rest } = input;
     return rest;
@@ -173,18 +164,18 @@ function getContent(input: any) {
   return interopRequireDefault(name);
 }
 
-export function loadPreset(
+export async function loadPreset(
   input: PresetConfig,
   level: number,
   storybookOptions: InterPresetOptions
-): LoadedPreset[] {
+): Promise<LoadedPreset[]> {
   try {
     // @ts-ignores
     const name: string = input.name ? input.name : input;
     // @ts-ignore
     const presetOptions = input.options ? input.options : {};
 
-    let contents = getContent(input);
+    let contents = await getContent(input);
 
     if (typeof contents === 'function') {
       // allow the export of a preset to be a function, that gets storybookOptions
@@ -203,12 +194,12 @@ export function loadPreset(
       const subAddons = resolvePresetFunction(addonsInput, presetOptions, storybookOptions);
 
       return [
-        ...loadPresets([...subPresets], level + 1, storybookOptions),
-        ...loadPresets(
+        ...(await loadPresets([...subPresets], level + 1, storybookOptions)),
+        ...(await loadPresets(
           [...subAddons.map(map(storybookOptions))].filter(Boolean) as PresetConfig[],
           level + 1,
           storybookOptions
-        ),
+        )),
         {
           name,
           preset: rest,
@@ -233,17 +224,22 @@ export function loadPreset(
   }
 }
 
-function loadPresets(
+async function loadPresets(
   presets: PresetConfig[],
   level: number,
   storybookOptions: InterPresetOptions
-): LoadedPreset[] {
+): Promise<LoadedPreset[]> {
   if (!presets || !Array.isArray(presets) || !presets.length) {
     return [];
   }
 
-  return presets.reduce<LoadedPreset[]>((acc, preset) => {
-    const loaded = loadPreset(preset, level, storybookOptions);
+  if (!level) {
+    logger.info('=> Loading presets');
+  }
+
+  return (
+    await Promise.all(presets.map(async (preset) => loadPreset(preset, level, storybookOptions)))
+  ).reduce((acc, loaded) => {
     return acc.concat(loaded);
   }, []);
 }
@@ -303,8 +299,11 @@ function applyPresets(
 
 type InterPresetOptions = Omit<CLIOptions & LoadOptions & BuilderOptions, 'frameworkPresets'>;
 
-export function getPresets(presets: PresetConfig[], storybookOptions: InterPresetOptions): Presets {
-  const loadedPresets: LoadedPreset[] = loadPresets(presets, 0, storybookOptions);
+export async function getPresets(
+  presets: PresetConfig[],
+  storybookOptions: InterPresetOptions
+): Promise<Presets> {
+  const loadedPresets: LoadedPreset[] = await loadPresets(presets, 0, storybookOptions);
 
   return {
     apply: async (extension: string, config: any, args = {}) =>
@@ -312,7 +311,7 @@ export function getPresets(presets: PresetConfig[], storybookOptions: InterPrese
   };
 }
 
-export function loadAllPresets(
+export async function loadAllPresets(
   options: CLIOptions &
     LoadOptions &
     BuilderOptions & {