+++ lib/core-common/src/types.ts
@@ -55,13 +55,13 @@ export interface Presets {
     config: TypescriptOptions,
     args?: Options
   ): Promise<TypescriptOptions>;
-  apply(extension: 'framework', config: {}, args: any): Promise<Preset>;
-  apply(extension: 'babel', config: {}, args: any): Promise<TransformOptions>;
-  apply(extension: 'entries', config: [], args: any): Promise<unknown>;
-  apply(extension: 'stories', config: [], args: any): Promise<StoriesEntry[]>;
-  apply(extension: 'managerEntries', config: [], args: any): Promise<string[]>;
-  apply(extension: 'refs', config: [], args: any): Promise<unknown>;
-  apply(extension: 'core', config: {}, args: any): Promise<CoreConfig>;
+  apply(extension: 'framework', config?: {}, args?: any): Promise<Preset>;
+  apply(extension: 'babel', config?: {}, args?: any): Promise<TransformOptions>;
+  apply(extension: 'entries', config?: [], args?: any): Promise<unknown>;
+  apply(extension: 'stories', config?: [], args?: any): Promise<StoriesEntry[]>;
+  apply(extension: 'managerEntries', config: [], args?: any): Promise<string[]>;
+  apply(extension: 'refs', config?: [], args?: any): Promise<unknown>;
+  apply(extension: 'core', config?: {}, args?: any): Promise<CoreConfig>;
   apply<T>(extension: string, config?: T, args?: unknown): Promise<T>;
 }
 
@@ -84,6 +84,7 @@ export interface Ref {
   title: string;
   version: string;
   type?: string;
+  disable?: boolean;
 }
 
 export interface VersionCheck {
@@ -129,11 +130,6 @@ export interface LoadOptions {
   configDir?: string;
   ignorePreview?: boolean;
   extendServer?: (server: Server) => void;
-}
-
-export interface ManagerWebpackOptions {
-  entries: string[];
-  refs: Record<string, Ref>;
 }
 
 export interface CLIOptions {
@@ -190,7 +186,7 @@ export interface StorybookConfigOptions {
 
 export type Options = LoadOptions & StorybookConfigOptions & CLIOptions & BuilderOptions;
 
-export interface Builder<Config, Stats> {
+export interface Builder<Config, BuilderStats extends Stats = Stats> {
   getConfig: (options: Options) => Promise<Config>;
   start: (args: {
     options: Options;
@@ -198,14 +194,14 @@ export interface Builder<Config, Stats> {
     router: Router;
     server: Server;
   }) => Promise<void | {
-    stats: Stats;
+    stats?: BuilderStats;
     totalTime: ReturnType<typeof process.hrtime>;
     bail: (e?: Error) => Promise<void>;
   }>;
   build: (arg: {
     options: Options;
     startTime: ReturnType<typeof process.hrtime>;
-  }) => Promise<void | Stats>;
+  }) => Promise<void | BuilderStats>;
   bail: (e?: Error) => Promise<void>;
   corePresets?: string[];
   overridePresets?: string[];
@@ -228,6 +224,7 @@ export interface StoryIndex {
 export interface StoryIndexer {
   test: RegExp;
   indexer: (fileName: string, options: IndexerOptions) => Promise<StoryIndex>;
+  addDocsTemplate?: boolean;
 }
 
 /**
@@ -289,6 +286,21 @@ export type Preset =
 export type Entry = string;
 
 type StorybookRefs = Record<string, { title: string; url: string } | { disable: boolean }>;
+
+export type DocsOptions = {
+  /**
+   * Should we generate docs entries at all under any circumstances? (i.e. can they be rendered)
+   */
+  enabled?: boolean;
+  /**
+   * What should we call the generated docs entries?
+   */
+  defaultName?: string;
+  /**
+   * Should we generate a docs entry per CSF file?
+   */
+  docsPage?: boolean;
+};
 
 /**
  * The interface for Storybook configuration in `main.ts` files.
@@ -417,6 +429,11 @@ export interface StorybookConfig {
    * Process CSF files for the story index.
    */
   storyIndexers?: (indexers: StoryIndexer[], options: Options) => StoryIndexer[];
+
+  /**
+   * Docs related features in index generation
+   */
+  docs?: DocsOptions;
 }
 
 export type PresetProperty<K, TStorybookConfig = StorybookConfig> =