+++ lib/core-server/src/build-dev.ts
@@ -72,7 +72,7 @@ export async function buildDevStandalone(options: CLIOptions & LoadOptions & Bui
   }
 
   logger.info('=> Loading presets');
-  let presets = loadAllPresets({
+  let presets = await loadAllPresets({
     corePresets,
     overridePresets: [],
     ...options,
@@ -80,7 +80,7 @@ export async function buildDevStandalone(options: CLIOptions & LoadOptions & Bui
 
   const [previewBuilder, managerBuilder] = await getBuilders({ ...options, presets });
 
-  presets = loadAllPresets({
+  presets = await loadAllPresets({
     corePresets: [
       require.resolve('./presets/common-preset'),
       ...managerBuilder.corePresets,
@@ -113,14 +113,12 @@ export async function buildDevStandalone(options: CLIOptions & LoadOptions & Bui
 
   if (options.webpackStatsJson) {
     const target = options.webpackStatsJson === true ? options.outputDir : options.webpackStatsJson;
-    await outputStats(target, previewStats, managerStats);
+    await outputStats(target, previewStats);
   }
 
   if (options.smokeTest) {
     const warnings: Error[] = [];
-    // @ts-ignore
     warnings.push(...((managerStats && managerStats.toJson().warnings) || []));
-    // @ts-ignore
     warnings.push(...((managerStats && previewStats.toJson().warnings) || []));
 
     const problems = warnings