+++ renderers/svelte/src/docs/extractArgTypes.ts
@@ -59,8 +59,9 @@ export const createArgTypes = (docgen: SvelteComponentDoc) => {
     docgen.events.forEach((item) => {
       results[`event_${item.name}`] = {
         name: item.name,
-        description: item.description || undefined,
-        type: { name: 'other', value: 'void' },
+        action: item.name,
+        control: false,
+        ...(item.description ? { description: item.description } : {}),
         table: {
           category: 'events',
         },
@@ -72,10 +73,10 @@ export const createArgTypes = (docgen: SvelteComponentDoc) => {
     docgen.slots.forEach((item) => {
       results[`slot_${item.name}`] = {
         name: item.name,
+        control: false,
         description: [item.description, item.params?.map((p) => `\`${p.name}\``).join(' ')]
           .filter((p) => p)
           .join('\n\n'),
-        type: { name: 'other', value: 'void' },
         table: {
           category: 'slots',
         },