+++ renderers/web-components/src/render.ts
@@ -2,7 +2,7 @@
 // @ts-ignore
 import global from 'global';
 
-import dedent from 'ts-dedent';
+import { dedent } from 'ts-dedent';
 import { render } from 'lit-html';
 // Keep `.js` extension to avoid issue with Webpack (related to export map?)
 // eslint-disable-next-line import/extensions